backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy>=1.26.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]==2.5.0
requests==2.31.0

